==> 15.gut_oral_comparison.Rmd <==
---
title: "gut_oral_comparison"
author: "Erica Ryu"
date: "9/13/2023"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 15. Gut-oral microbiome comparison

The purpose of this script is to compare the gut and oral microbiomes

## load packages
```{r}
library(phyloseq)
library(vegan)
library(ggplot2)
library(data.table)
library(dplyr)
library(tidyr)
library(magrittr)
library(MASS)
library(clinfun)
```

## load data
```{r}
# load in phyloseq object
phyloseq_complete <- readRDS("output/ps_complete.rds")

## subset based on extraction kit
qiagen <- subset_samples(phyloseq_complete, Condition == "Qiagen")

# gut microbiome
gut <- readRDS("data/gut_phyloseq.rds")
```

## set up functions
```{r}
# calculate beta diversity
beta_ordinate <- function(physeq){
  # set up data
  ps.prop <- transform_sample_counts(physeq, function(otu) otu/sum(otu))
  ps.prop@otu_table <- na.omit(ps.prop@otu_table)
  ## calculate distance and ordinate
  ord.pcoa <- ordinate(ps.prop, method = "PCoA", distance = "bray")
}

# set up beta diversity data for plotting
plot_beta <- function(beta, physeq){
  # extract PCoA axes
  PCOAaxes <- beta$vectors[,c(1,2,3,4)]
  # extract lifestyle column from metadata and add to vectors
  lifestyle <- physeq@sam_data$Lifestyle
  PCOAaxes_meta <- cbind(PCOAaxes, lifestyle)
  df_PCOA <- as.data.frame(as.matrix(PCOAaxes_meta))
  # add space to Recently Settled
  df_PCOA$lifestyle <- gsub("RecentlySettled", "Recently Settled", df_PCOA$lifestyle)
  # set lifestyle stages to levels
  df_PCOA$lifestyle <- factor(df_PCOA$lifestyle, levels=c("Foragers", "Recently Settled", "Agriculturalists", "Expats"))
  return(df_PCOA)
}
```

## subset to sample IDs
```{r}
qiagen_metadata <- qiagen@sam_data
qiagen_metadata <- as.data.frame(as.matrix(qiagen_metadata))
oral_samples <- qiagen_metadata$Description

gut_metadata <- gut@sam_data
gut_metadata <- as.data.frame(as.matrix(gut_metadata))
gut_samples <- gut_metadata$SampleID
```

## sort samples based on whether they are in both, only gut, or only oral
```{r}
gut_oral <- c(oral_samples, gut_samples)
# remove duplicates
gut_oral <- unique(gut_oral)

both <- c()
gut_only <- c()
oral_only <- c()

# sort samples into vectors
for (i in gut_oral) {
  if (i %in% gut_samples & i %in% oral_samples){
    both <- c(both, i)
  } else if (i %in% gut_samples & !(i %in% oral_samples)){
      gut_only <- c(gut_only, i)
  } else if (i %in% oral_samples & !(i %in% gut_samples)) {
      oral_only <- c(oral_only, i)
    } else print(i)
}

# make vectors the same length
n <- max(length(both), length(gut_only), length(oral_only))
length(both) <- n
length(gut_only) <- n
length(oral_only) <- n

overlap <- as.data.frame(cbind(both, gut_only, oral_only))
overlap_names <- paste0(overlap$both, "_genotek")

write.csv(overlap, row.names = FALSE, na = "", file = "output/gut_oral_overlap.csv")
```

## filter to samples present in both microbiomes
```{r}
# subset phyloseq objects to overlapping samples
qiagen_overlap <- subset_samples(qiagen, SampleID %in% overlap_names)
gut_overlap <- subset_samples(gut, SampleID %in% overlap$both)
```

## generate PCoA axes
```{r}
# generate PCoA data for oral microbiome
bray_ordinate_qia <- beta_ordinate(qiagen_overlap)
bray_plot_qia <- plot_beta(bray_ordinate_qia, qiagen_overlap)
bray_plot_num_qia <-  as.data.frame(sapply(bray_plot_qia, as.numeric))
rownames(bray_plot_num_qia) <- rownames(bray_plot_qia)
colnames(bray_plot_num_qia) <- c("PCoA1_oral", "PCoA2_oral", "PCoA3_oral", "PCoA4_oral", "lifestyle")
bray_plot_num_qia$lifestyle <- ifelse(grepl("CHE", rownames(bray_plot_num_qia)), "Foragers", 
                              ifelse(grepl("THA", rownames(bray_plot_num_qia)), "Agriculturalists", "Recently Settled"))
bray_plot_num_qia$lifestyle <- factor(bray_plot_num_qia$lifestyle, levels=c("Foragers", "Recently Settled", "Agriculturalists"))
dim(bray_plot_num_qia)
#$ remove _genotek from rownames
rownames(bray_plot_num_qia) <- gsub("_.*", "", rownames(bray_plot_num_qia))

# generate PCoA data for gut microbiome
bray_ordinate_gut <- beta_ordinate(gut_overlap)
PCOAaxes_gut <- bray_ordinate_gut$vectors[,c(1,2,3,4)]
df_PCOA_gut <- as.data.frame(as.matrix(PCOAaxes_gut))
bray_plot_num_gut <-  as.data.frame(sapply(df_PCOA_gut, as.numeric))
rownames(bray_plot_num_gut) <- rownames(df_PCOA_gut)
colnames(bray_plot_num_gut) <- c("PCoA1_gut", "PCoA2_gut", "PCoA3_gut", "PCoA4_gut")
bray_plot_num_gut$lifestyle <- ifelse(grepl("CHE", rownames(bray_plot_num_gut)), "Foragers", 
                              ifelse(grepl("THA", rownames(bray_plot_num_gut)), "Agriculturalists", "Recently Settled"))
bray_plot_num_gut$lifestyle <- factor(bray_plot_num_gut$lifestyle, levels=c("Foragers", "Recently Settled", "Agriculturalists"))
dim(bray_plot_num_gut)

# combine into one table
pcoa_comb <- merge(bray_plot_num_qia, bray_plot_num_gut, by = 0)

# clean table
rownames(pcoa_comb) <- pcoa_comb$Row.names
pcoa_comb <- dplyr::select(pcoa_comb, -c("Row.names", "lifestyle.x"))
```

## plot PCoA axes
```{r}
# PCoA1 vs PCoA1
plot_overlap_oralgut_pcoa1 <- ggplot(pcoa_comb, aes(x = PCoA1_oral, y = PCoA1_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA1 oral vs PCoA1 gut",
       x = "PCoA1 oral",
       y = "PCoA1 gut")

## correlation
cor.test(pcoa_comb$PCoA1_oral, pcoa_comb$PCoA1_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralgut_pcoa1.pdf", plot=plot_overlap_oralgut_pcoa1, width=4, height=2)

# PCoA1 vs PCoA2
plot_overlap_oralpcoa1_gutpcoa2 <- ggplot(pcoa_comb, aes(x = PCoA1_oral, y = PCoA2_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA1 oral vs PCoA2 gut",
       x = "PCoA1 oral",
       y = "PCoA2 gut")

## correlation
cor.test(pcoa_comb$PCoA1_oral, pcoa_comb$PCoA2_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralpcoa1_gutpcoa2.pdf", plot=plot_overlap_oralpcoa1_gutpcoa2, width=4, height=2)

# PCoA2 vs PCoA1
plot_overlap_oralpcoa2_gutpcoa1 <- ggplot(pcoa_comb, aes(x = PCoA2_oral, y = PCoA1_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA2 oral vs PCoA1 gut",
       x = "PCoA2 oral",
       y = "PCoA1 gut")

## correlation
cor.test(pcoa_comb$PCoA2_oral, pcoa_comb$PCoA1_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralpcoa2_gutpcoa1.pdf", plot=plot_overlap_oralpcoa2_gutpcoa1, width=4, height=2)

# PCoA2 vs PCoA2
plot_overlap_oralgut_pcoa2 <- ggplot(pcoa_comb, aes(x = PCoA2_oral, y = PCoA2_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA2 oral vs PCoA2 gut",
       x = "PCoA2 oral",
       y = "PCoA2 gut")

## correlation
cor.test(pcoa_comb$PCoA2_oral, pcoa_comb$PCoA2_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralgut_pcoa2.pdf", plot=plot_overlap_oralgut_pcoa2, width=4, height=2)

# PCoA1 vs PCoA3
plot_overlap_oralpcoa1_gutpcoa3 <- ggplot(pcoa_comb, aes(x = PCoA1_oral, y = PCoA3_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA1 oral vs PCoA3 gut",
       x = "PCoA1 oral",
       y = "PCoA3 gut")

## correlation
cor.test(pcoa_comb$PCoA1_oral, pcoa_comb$PCoA3_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralpcoa1_gutpcoa3.pdf", plot=plot_overlap_oralpcoa1_gutpcoa3, width=4, height=2)

# PCoA2 vs PCoA3
plot_overlap_oralpcoa2_gutpcoa3 <- ggplot(pcoa_comb, aes(x = PCoA2_oral, y = PCoA3_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(x = "PCoA2 oral",
       y = "PCoA3 gut")

## correlation
cor.test(pcoa_comb$PCoA2_oral, pcoa_comb$PCoA3_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralpcoa2_gutpcoa3.pdf", plot=plot_overlap_oralpcoa2_gutpcoa3, width=4, height=2)

# PCoA3 vs PCoA1
plot_overlap_oralpcoa3_gutpcoa1 <- ggplot(pcoa_comb, aes(x = PCoA3_oral, y = PCoA1_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA3 oral vs PCoA1 gut",
       x = "PCoA3 oral",
       y = "PCoA1 gut")

## correlation
cor.test(pcoa_comb$PCoA3_oral, pcoa_comb$PCoA1_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralpcoa3_gutpcoa1.pdf", plot=plot_overlap_oralpcoa3_gutpcoa1, width=4, height=2)

# PCoA3 vs PCoA2
plot_overlap_oralpcoa3_gutpcoa2 <- ggplot(pcoa_comb, aes(x = PCoA3_oral, y = PCoA2_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA3 oral vs PCoA2 gut",
       x = "PCoA3 oral",
       y = "PCoA2 gut")

## correlation
cor.test(pcoa_comb$PCoA3_oral, pcoa_comb$PCoA2_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralpcoa3_gutpcoa2.pdf", plot=plot_overlap_oralpcoa3_gutpcoa2, width=4, height=2)

# PCoA3 vs PCoA3
plot_overlap_oralgut_pcoa3 <- ggplot(pcoa_comb, aes(x = PCoA3_oral, y = PCoA3_gut)) +
  geom_point(shape = 21, color = "black", size = 3, aes(fill = lifestyle.y)) +
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_smooth(method = "glm") +
  labs(title = "PCoA3 oral vs PCoA3 gut",
       x = "PCoA3 oral",
       y = "PCoA3 gut")

#$ correlation
cor.test(pcoa_comb$PCoA3_oral, pcoa_comb$PCoA3_gut, alternative = "two.sided", method=c("spearman"))

ggsave(file="figures/plot_overlap_oralgut_pcoa3.pdf", plot=plot_overlap_oralgut_pcoa3, width=4, height=2)
```

## prep gut and oral ASV tables
```{r}
qiagen_overlap_relabund <- transform_sample_counts(qiagen_overlap, function(otu) otu/sum(otu))
gut_overlap_relabund <- transform_sample_counts(gut_overlap, function(otu) otu/sum(otu))

# convert to dfs
qiagen_overlap_df <- as.data.frame(qiagen_overlap_relabund@otu_table)
gut_overlap_df <- as.data.frame(gut_overlap_relabund@otu_table)

# check for intersection in ASVs
intersect(colnames(qiagen_overlap_df), colnames(gut_overlap_df))

# combine
qiagen_gut_comb <- bind_rows(qiagen_overlap_df, gut_overlap_df)

# replace NAs with 0
qiagen_gut_comb[is.na(qiagen_gut_comb)] <- 0

# calculate bray curtis distance
qiagen_gut_comb_dist <- vegdist(qiagen_gut_comb, method = "bray")

# format
qiagen_gut_comb_dist <- as.matrix(qiagen_gut_comb_dist)
```

## prep genera tables for distance calculation
```{r}
phy_genus <- tax_glom(qiagen_overlap, "Genus", NArm = FALSE);phy_genus
phy_genus_RA <- transform_sample_counts(phy_genus, function(otu) otu/sum(otu))
otu_table <- as.data.frame(as.matrix(phy_genus_RA@otu_table))

setDT(otu_table, keep.rownames = TRUE)[]
colnames(otu_table)[1] <- "SampleID"

otu_table2 <- otu_table %>% data.frame %>% magrittr::set_rownames(.$SampleID)
otu_table2 <- otu_table2[,-1]
otu_table2 <- t(otu_table2)
tax_table <- as.data.frame(as.matrix(phy_genus@tax_table))

otu_tax <- merge(otu_table2, tax_table, by = 0, all = TRUE)

# change colnames to genus name
otu_tax$Genus <- replace_na(otu_tax$Genus, "_unclassified")
# and add higher taxonomic level to unclassified
otu_tax$Phylum <- ifelse(is.na(otu_tax$Phylum), otu_tax$Kingdom, otu_tax$Phylum)
otu_tax$Class <- ifelse(is.na(otu_tax$Class), otu_tax$Phylum, otu_tax$Class)
otu_tax$Order <- ifelse(is.na(otu_tax$Order), otu_tax$Class, otu_tax$Order)
otu_tax$Family <- ifelse(is.na(otu_tax$Family), otu_tax$Order, otu_tax$Family)

otu_tax$Genus <- ifelse(otu_tax$Genus == "_unclassified", paste(otu_tax$Family, otu_tax$Genus, sep = ""), otu_tax$Genus)

# rename completely unclassified ASV
otu_tax$Genus <- gsub("NA_unclassified", "Unclassified", otu_tax$Genus)

# make genus rownames
rownames(otu_tax) <- otu_tax$Genus
# and remove taxonomic columns
otu_tax <- subset(otu_tax, select = -c(Row.names, Kingdom, Phylum, Class, Order, Family, Genus))
# transpose
otu_tax_t <- t(otu_tax)
otu_tax_t <- as.data.frame(otu_tax_t)

### GUT
phy_genus_gut <-  tax_glom(gut_overlap, "Genus", NArm = FALSE);phy_genus_gut
phy_genus_gut_RA <- transform_sample_counts(phy_genus_gut, function(otu) otu/sum(otu))
otu_table_gut <- as.data.frame(as.matrix(phy_genus_gut_RA@otu_table))

setDT(otu_table_gut, keep.rownames = TRUE)[]
colnames(otu_table_gut)[1] <- "SampleID"

otu_table2_gut <- otu_table_gut %>% data.frame %>% set_rownames(.$SampleID)
otu_table2_gut <- otu_table2_gut[,-1]
otu_table2_gut <- t(otu_table2_gut)
tax_table_gut <- as.data.frame(as.matrix(phy_genus_gut@tax_table))

otu_tax_gut <- merge(otu_table2_gut, tax_table_gut, by = 0, all = TRUE)

# change colnames to genus name
otu_tax_gut$Genus <- replace_na(otu_tax_gut$Genus, "_unclassified")
# and add higher taxonomic level to unclassified
otu_tax_gut$Phylum <- ifelse(is.na(otu_tax_gut$Phylum), otu_tax_gut$Kingdom, otu_tax_gut$Phylum)
otu_tax_gut$Class <- ifelse(is.na(otu_tax_gut$Class), otu_tax_gut$Phylum, otu_tax_gut$Class)
otu_tax_gut$Order <- ifelse(is.na(otu_tax_gut$Order), otu_tax_gut$Class, otu_tax_gut$Order)
otu_tax_gut$Family <- ifelse(is.na(otu_tax_gut$Family), otu_tax_gut$Order, otu_tax_gut$Family)

otu_tax_gut$Genus <- ifelse(otu_tax_gut$Genus == "_unclassified", paste(otu_tax_gut$Family, otu_tax_gut$Genus, sep = ""), otu_tax_gut$Genus)

# rename completely unclassified ASV
otu_tax_gut$Genus <- gsub("NA_unclassified", "Unclassified", otu_tax_gut$Genus)

# make genus rownames
rownames(otu_tax_gut) <- otu_tax_gut$Genus
# and remove taxonomic columns
otu_tax_gut <- subset(otu_tax_gut, select = -c(Row.names, Kingdom, Phylum, Class, Order, Family, Genus))
# transpose
otu_tax_gut_t <- t(otu_tax_gut)
otu_tax_gut_t <- as.data.frame(otu_tax_gut_t)
```

## calculate distance between oral and gut microbiomes
```{r}
# check for intersection in genus
intersect(colnames(otu_tax_t), colnames(otu_tax_gut_t))

# combine
qiagen_gut_genus_comb <- bind_rows(otu_tax_t, otu_tax_gut_t)

# replace NAs with 0
qiagen_gut_genus_comb[is.na(qiagen_gut_genus_comb)] <- 0

# calculate bray curtis distance
qiagen_gut_genus_comb_dist <- vegdist(qiagen_gut_genus_comb, method = "bray")

# format
qiagen_gut_genus_comb_dist <- as.matrix(qiagen_gut_genus_comb_dist)

# filter to unique columns and rows
qiagen_gut_genus_comb_dist_filt <- qiagen_gut_genus_comb_dist[grepl("_genotek", rownames(qiagen_gut_genus_comb_dist)),]

qiagen_gut_genus_comb_dist_filt <- qiagen_gut_genus_comb_dist_filt[,!grepl("_genotek", colnames(qiagen_gut_genus_comb_dist_filt))]

# get distance across same individuals
qia_gut_indiv_dist <- diag(qiagen_gut_genus_comb_dist_filt)

qia_gut_indiv_dist_tab <- data.frame(colnames(qiagen_gut_genus_comb_dist_filt), qia_gut_indiv_dist)

# add colnames
colnames(qia_gut_indiv_dist_tab) <- c("sample", "distance")

# add lifestyle info
qia_gut_indiv_dist_tab$lifestyle <- ifelse(grepl("CHE", qia_gut_indiv_dist_tab$sample), "Foragers", 
                              ifelse(grepl("THA", qia_gut_indiv_dist_tab$sample), "Agriculturalists", 
                                     "Recently Settled"))

qia_gut_indiv_dist_tab$lifestyle <- factor(qia_gut_indiv_dist_tab$lifestyle, ordered = TRUE, levels=c("Foragers", "Recently Settled", "Agriculturalists"))

# plot
overlap_beta <- ggplot(qia_gut_indiv_dist_tab, aes(x = lifestyle, y = distance)) +
  geom_violin(aes(fill=lifestyle), alpha=0.8) +
  geom_boxplot(width=0.1, color="black", alpha=0.2, outlier.shape = NA) +
  scale_x_discrete(limits=c("Foragers", "Recently Settled", "Agriculturalists")) + 
  scale_fill_manual(name=NULL,
                    values=c("darkslateblue", "deepskyblue", "lightblue3"),
                    breaks=c("Foragers", "Recently Settled", "Agriculturalists"),
                    labels=c("Foragers", "Recently Settled", "Agriculturalists")) +
  geom_jitter(size = 1, col="darkgreen", width = 0.1) + 
  labs(x = "Lifestyle", y = "Bray-Curtis Distance")+
  theme(axis.text.x = element_text(angle = 0, hjust = 0.5, vjust=0.5)) + 
  theme(legend.position = "none")

library(DescTools)
JonckheereTerpstraTest(distance ~ lifestyle, data = qia_gut_indiv_dist_tab)

ggsave(file="figures/overlap_beta.pdf", plot=overlap_beta, width=6, height=3)
```

## power calculation
```{r}
# get sample count
forager_count <- length(which(qia_gut_indiv_dist_tab$lifestyle == "Foragers"))
rs_count <- length(which(qia_gut_indiv_dist_tab$lifestyle == "Recently Settled"))
ag_count <- length(which(qia_gut_indiv_dist_tab$lifestyle == "Agriculturalists"))

# get effect sizes
df_gut_oral_dist <- as.data.frame(qia_gut_indiv_dist_tab)
df_gut_oral_dist <- subset(df_gut_oral_dist, select=-c(sample))
df_gut_oral_dist_forg_rs <- subset(df_gut_oral_dist, lifestyle != "Agriculturalists")
df_gut_oral_dist_forg_ag <- subset(df_gut_oral_dist, lifestyle != "Recently Settled")

## forg - rs
forg_rs_mean <- df_gut_oral_dist_forg_rs %>%
    # Specify group indicator, column, function
    group_by(lifestyle) %>%
    # Calculate the mean of the "Frequency" column for each group
    summarise_at(vars(distance),
                 list(Mean_Frequency = mean))

forg_rs_sd <- df_gut_oral_dist_forg_rs %>%
    # Specify group indicator, column, function
    group_by(lifestyle) %>%
    # Calculate the mean of the "Frequency" column for each group
    summarise_at(vars(distance),
                 list(sd_Frequency = sd))

forg_mean <- 0.8290034
rs_mean <- 0.8221369
forg_sd <- 0.1453378
rs_sd <- 0.1432843

pool_forg_rs_sd <- sqrt((((rs_count - 1) * rs_sd^2) + ((forager_count - 1) * forg_sd^2))/(forager_count + rs_count - 2))

d_forg_rs <- (forg_mean - rs_mean)/pool_forg_rs_sd

## rs-ag
rs_ag_mean <- df_gut_oral_dist_rs_ag %>%
    # Specify group indicator, column, function
    group_by(lifestyle) %>%
    # Calculate the mean of the "Frequency" column for each group
    summarise_at(vars(distance),
                 list(Mean_Frequency = mean))

rs_ag_sd <- df_gut_oral_dist_rs_ag %>%
    # Specify group indicator, column, function
    group_by(lifestyle) %>%
    # Calculate the mean of the "Frequency" column for each group
    summarise_at(vars(distance),
                 list(sd_Frequency = sd))

forg_mean <- 0.829003
ag_mean <- 0.7762685
forg_sd <- 0.1453378
ag_sd <- 0.09477919

pool_forg_ag_sd <- sqrt((((ag_count - 1) * ag_sd^2) + ((forager_count - 1) * forg_sd^2))/(forager_count + ag_count - 2))

d_forg_ag <- (forg_mean - ag_mean)/pool_forg_ag_sd

# calculate
power.ladesign(c(12, 14, 12), odds.ratio = c(d_forg_rs, d_forg_ag), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))
```

## power analysis
```{r}
# 20 
forager_20 <- 20
rs_20 <- 20
ag_20 <- 20

pool_forg_rs_sd_20 <- sqrt((((rs_20 - 1) * rs_sd^2) + ((forager_20 - 1) * forg_sd^2))/(forager_20 + rs_20 - 2))

d_forg_rs_20 <- (forg_mean - rs_mean)/pool_forg_rs_sd_20

pool_forg_ag_sd_20 <- sqrt((((ag_20 - 1) * ag_sd^2) + ((forager_20 - 1) * forg_sd^2))/(forager_20 + ag_20 - 2))

d_forg_ag_20 <- (forg_mean - ag_mean)/pool_forg_ag_sd_20

power.ladesign(c(20, 20, 20), odds.ratio = c(d_forg_rs_20, d_forg_ag_20), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))


# 30 
forager_30 <- 30
rs_30 <- 30
ag_30 <- 30

pool_forg_rs_sd_30 <- sqrt((((rs_30 - 1) * rs_sd^2) + ((forager_30 - 1) * forg_sd^2))/(forager_30 + rs_30 - 2))

d_forg_rs_30 <- (forg_mean - rs_mean)/pool_forg_rs_sd_30

pool_forg_ag_sd_30 <- sqrt((((ag_30 - 1) * ag_sd^2) + ((forager_30 - 1) * forg_sd^2))/(forager_30 + ag_30 - 2))

d_forg_ag_30 <- (forg_mean - ag_mean)/pool_forg_ag_sd_30

power.ladesign(c(30, 30, 30), odds.ratio = c(d_forg_rs_30, d_forg_ag_30), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))


# 40 
forager_40 <- 40
rs_40 <- 40
ag_40 <- 40

pool_forg_rs_sd_40 <- sqrt((((rs_40 - 1) * rs_sd^2) + ((forager_40 - 1) * forg_sd^2))/(forager_40 + rs_40 - 2))

d_forg_rs_40 <- (forg_mean - rs_mean)/pool_forg_rs_sd_40

pool_forg_ag_sd_40 <- sqrt((((ag_40 - 1) * ag_sd^2) + ((forager_40 - 1) * forg_sd^2))/(forager_40 + ag_40 - 2))

d_forg_ag_40 <- (forg_mean - ag_mean)/pool_forg_ag_sd_40

power.ladesign(c(40, 40, 40), odds.ratio = c(d_forg_rs_40, d_forg_ag_40), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))


# 50 
forager_50 <- 50
rs_50 <- 50
ag_50 <- 50

pool_forg_rs_sd_50 <- sqrt((((rs_50 - 1) * rs_sd^2) + ((forager_50 - 1) * forg_sd^2))/(forager_50 + rs_50 - 2))

d_forg_rs_50 <- (forg_mean - rs_mean)/pool_forg_rs_sd_50

pool_forg_ag_sd_50 <- sqrt((((ag_50 - 1) * ag_sd^2) + ((forager_50 - 1) * forg_sd^2))/(forager_50 + ag_50 - 2))

d_forg_ag_50 <- (forg_mean - ag_mean)/pool_forg_ag_sd_50

power.ladesign(c(50, 50, 50), odds.ratio = c(d_forg_rs_50, d_forg_ag_50), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))


# 60 
forager_60 <- 60
rs_60 <- 60
ag_60 <- 60

pool_forg_rs_sd_60 <- sqrt((((rs_60 - 1) * rs_sd^2) + ((forager_60 - 1) * forg_sd^2))/(forager_60 + rs_60 - 2))

d_forg_rs_60 <- (forg_mean - rs_mean)/pool_forg_rs_sd_60

pool_forg_ag_sd_60 <- sqrt((((ag_60 - 1) * ag_sd^2) + ((forager_60 - 1) * forg_sd^2))/(forager_60 + ag_60 - 2))

d_forg_ag_60 <- (forg_mean - ag_mean)/pool_forg_ag_sd_60

power.ladesign(c(60, 60, 60), odds.ratio = c(d_forg_rs_60, d_forg_ag_60), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))


# 62 
forager_62 <- 62
rs_62 <- 62
ag_62 <- 62

pool_forg_rs_sd_62 <- sqrt((((rs_62 - 1) * rs_sd^2) + ((forager_62 - 1) * forg_sd^2))/(forager_62 + rs_62 - 2))

d_forg_rs_62 <- (forg_mean - rs_mean)/pool_forg_rs_sd_62

pool_forg_ag_sd_62 <- sqrt((((ag_62 - 1) * ag_sd^2) + ((forager_62 - 1) * forg_sd^2))/(forager_62 + ag_62 - 2))

d_forg_ag_62 <- (forg_mean - ag_mean)/pool_forg_ag_sd_62

power.ladesign(c(62, 62, 62), odds.ratio = c(d_forg_rs_62, d_forg_ag_62), sig.level = 0.05, statistic = c("Jonckheere-Terpstra"), alternative = c("two.sided"))
```

## plot power analysis
```{r}
samp_size <- c(12.7, 20, 30, 40, 50, 60, 62)
power_lev <- c(0.041, 0.156, 0.327, 0.502, 0.66, 0.785, 0.804)
df_power_plot <- as.data.frame(cbind(samp_size, power_lev))

power_plot <- ggplot(df_power_plot, aes(x = samp_size, y = power_lev)) +
  geom_line() +
  geom_hline(yintercept = 0.8, linetype="dashed", color = "blue") +
  geom_hline(yintercept = 0.041, linetype="dashed", color = "red") +
  geom_point(shape = 16, color = "black", size = 3)  + 
  labs(x = "Sample Size", y = "Predicted Power") +
  theme(legend.position = "none")

ggsave(file="figures/power_plot.pdf", plot=power_plot, width=6, height=3)

```
